Modularized Header, created separate components.
